+++ src/migrations/Install.php
@@ -1,5 +1,7 @@
 <?php
 
+declare(strict_types=1);
+
 namespace kernpfad\cleverreach\migrations;
 
 use craft\db\Migration;